LLM_FN_Analysis/requirements.txt
pandas>=1.5.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.28.0
numpy>=1.21.0
pathlib2>=2.3.0; python_version < "3.4"
